0.2.1
Updated /config/connection.js to have console.log that notifies the db was sucesfully connected, updated /models/user.js to have virtuals and to utilize the bcrypt package so that sensitive information would be hashed prior to creation of an instance of the user. Updated /model/project.js to have a projectId field, title, and pass in the tasks schema to projects tasks key inside of projects, Created a /models/tasks.js that will refrence the projectId, have a task sate, taskbody, difficulty and task owner. Uncommented db connection related code inside of server.js.