0.9.1
Fixed navbar so that auth/utils function getToken validates if a user is logged in or not and changes visual display and options of Navbar function component, added a onlick event to logout button that uses a callback to initalize auth.logout which destroys the saved token. Fixed bug where models wouldnt open on mobile.