0.11.0
Client: bugixed addprojects by making the state not reassign which for some reason cause 400. update current project so the key is set to the project._id was project.id mistakenly. commented out the completed projects header and the completedprojects component as that is a future feature and not a MVP participantupdated utils/mutations.js to have a functional DEL_PROJECT graphql method, Server: updated graphql/resolvers to use the where _id: is equal to args._id of the project since, projectId doesnt have data.