0.14.0
CLIENT: updated currentprojects.jsx component to have beeter indentation, updated the way the href and id are put into the url it is now not an endpoint but instead a query string. so that you can hit the /projects endpoint with the id within the url still, updated page/project.jsx to get the window.location.href and split the url based on the = and assign that id to a variable that will be used to pass in as an arguemnt for the return_project query resolver, sucessfully returned data and dynamically rendered project based off of the query string from the url. updated utils/query.js added RETURN_PROJECT GQL snippit to return all wanted data. SERVER: updated typedefs to have the new query method return_project for project.jsx, added new resolver async function to return the projects from the usermodel since the project is a embbeddoc used the .find method inside of the resolver and returned selectedProject variable, to populate with specific data. Conclusion: You can now while logged in, navigate to the dashboard add a project, click on a project and be directed to an instance of the projects.jsx page with dynamically generated data from the id that was input to the url as a querystring, based off the id of which project your clicked.