0.19.1
BUGFIXES: fixed the logic inside of project.js to instead of using multiple reduce methods uses only one and allows for empty feilds prior 0 items in a feild would cause error where you could not reduce over empty arr,  changed project column icon to have a trashcan instead of a pencil icon as it is intended to be a delete the task. updated current projects.jsx to remove status feild from project as it is always null and decided against setting an overall project status. updated critternameform so that after completion it will redirect to the same page to close module and show reflected changes in the name of the critter.